+++ src/features/Admin/components/Trash/ListTrasp.jsx
@@ -11,7 +11,16 @@ import { useDispatch, useSelector } from 'react-redux';
 import { useHistory, useRouteMatch } from 'react-router';
 
 
+import AcUnitIcon from '@material-ui/icons/AcUnit';
+import AddCircleIcon from '@material-ui/icons/AddCircle';
+import { useSnackbar } from 'notistack';
+import Swal from 'sweetalert2';
+
+
+
 function ListTrash(props) {
+
+  const { enqueueSnackbar } = useSnackbar()
 
   const {
     params: { idPart },
@@ -24,18 +33,45 @@ function ListTrash(props) {
 
   const handleRestoreTopic = async (idTrash) => {
     const action = restoreTopic(idTrash)
-    const rsRestore = await dispath(action)
-    
+    const resultAction = await dispath(action)
+
+    let slipType = resultAction.type.split('/')
+
+    if (slipType[2] == 'fulfilled') {
+
+      enqueueSnackbar('Restore Successfully', { variant: 'success' })
+    } else {
+      enqueueSnackbar('Restore Not Successfully', { variant: 'error' })
+    }
   }
 
   const handlDeleteTopic = async (idTrash) => {
-    const action = deleteTopic(idTrash)
-    const rsRestore = await dispath(action)
-    
-  }
 
-  const handleAddTopic = (a) => {
-    history.push(`/admin/parts/${a}/edit`);
+
+    Swal.fire({
+      title: 'Are you sure?',
+      text: "You won't be able to revert this!",
+      icon: 'warning',
+      showCancelButton: true,
+      confirmButtonColor: '#1976d2',
+      cancelButtonColor: '#dc004e',
+      confirmButtonText: 'Yes, delete it!'
+    }).then((result) => {
+      if (result.isConfirmed) {
+        (async () => {
+
+          const action = deleteTopic(idTrash)
+          const resultAction = await dispath(action)
+
+          let slipType = resultAction.type.split('/')
+          if (slipType[2] == 'fulfilled') {
+            enqueueSnackbar('Delete Successfully', { variant: 'success' })
+          } else {
+            enqueueSnackbar('Delete Not Successfully', { variant: 'error' })
+          }
+        })(); 
+      }
+    })
   }
 
   useEffect(() => {
@@ -46,52 +82,55 @@ function ListTrash(props) {
   }, []);
   let i = 0;
   return (
-    <div>
-      <Box>
-        <Grid container>
-          <Grid item lg={12} >
-            <Paper>List Trash</Paper>
-          </Grid>
-          <Grid item lg={12}>
-            <table class="content-table">
-              <thead>
-                <tr>
-                  <th>#</th>
-                  <th>IMG</th>
-                  <th>Name</th>
-                  <th>Description</th>
-                  <th>Action</th>
-                </tr>
-              </thead>
-              <tbody>
+    <Box className='content-list'>
+      <Grid container>
+        <Grid item lg={12}>
+          <Box className='ad-table'>
 
-                {trashState.map((topic) => (
+            <Box className='table-title'>
+              <Box className='content-title'>
+                <AcUnitIcon className='table-icon' />
+                <Typography variant='h5' className='part-text'> This Is My Gu</Typography>
+              </Box>
+            </Box>
 
-                  <tr key={topic._id}>
-                    <td>{i++}</td>
-                    <td>{topic.nameTopic}</td>
-                    <td> {topic.img}</td>
-                    <td>{topic.descripTopic}</td>
-                    <td>
-                      <Button onClick={() => handleRestoreTopic(topic._id)}>Restore</Button>
-                      <Button onClick={() => handlDeleteTopic(topic._id)}>Delete</Button>
+            <Box className='table-body'>
 
-                    </td>
+              <table class="content-table">
+                <thead>
+                  <tr>
+                    <th>#</th>
+                    <th>IMG</th>
+                    <th>Name</th>
+                    <th>Description</th>
+                    <th>Action</th>
                   </tr>
-                ))}
+                </thead>
+                <tbody>
 
-              </tbody>
-            </table>
-          </Grid>
+                  {trashState.map((topic) => (
 
-          <Grid item lg={12} >
-            <Button onClick={() => handleAddTopic()}>Edit</Button>
-          </Grid>
+                    <tr key={topic._id}>
+                      <td>{i++}</td>
+                      <td>{topic.nameTopic}</td>
+                      <td> {topic.img}</td>
+                      <td>{topic.descripTopic}</td>
+                      <td>
+                        <Button onClick={() => handleRestoreTopic(topic._id)}>Restore</Button>
+                        <Button onClick={() => handlDeleteTopic(topic._id)}>Delete</Button>
+
+                      </td>
+                    </tr>
+                  ))}
+
+                </tbody>
+              </table>
+
+            </Box>
+          </Box>
         </Grid>
-
-
-      </Box>
-    </div>
+      </Grid>
+    </Box>
   );
 }
